updated folder structure
